Move Repository and DataStorage to persistence
Both of these classes deal with persistence and it makes more sense to
have those in a common place.
