+++ src/LibreCards.Core/Persistence/ICardRepository.cs
@@ -1,7 +1,7 @@
-using System.Collections.Generic;
 using LibreCards.Core.Entities;
+using System.Collections.Generic;
 
-namespace LibreCards.Core
+namespace LibreCards.Core.Persistence
 {
     public interface ICardRepository
     {
